dashboard/requirements.txt
dash
dash_daq
dash-bootstrap-components
pandas
plotly
mysql-connector-python
wordcloud